+++ src/widgets.js
@@ -88,7 +88,7 @@ ScrollFrameMorph, MenuItemMorph, useBlurredShadows, getDocumentPositionOf*/
 
 /*jshint esversion: 6*/
 
-modules.widgets = '2025-December-09';
+modules.widgets = '2025-December-17';
 
 var PushButtonMorph;
 var ToggleButtonMorph;
@@ -278,7 +278,7 @@ PushButtonMorph.prototype.drawOutline = function (ctx) {
     ctx.beginPath();
     this.outlinePath(
         ctx,
-        isFlat ? 0 : this.corner,
+        this.corner,
         0
     );
     ctx.closePath();
@@ -286,13 +286,11 @@ PushButtonMorph.prototype.drawOutline = function (ctx) {
 };
 
 PushButtonMorph.prototype.drawBackground = function (ctx, color) {
-    var isFlat = MorphicPreferences.isFlat && !this.is3D;
-
     ctx.fillStyle = color.toString();
     ctx.beginPath();
     this.outlinePath(
         ctx,
-        isFlat ? 0 : Math.max(this.corner - this.outline, 0),
+        Math.max(this.corner - this.outline, 0),
         this.outline
     );
     ctx.closePath();
@@ -2935,7 +2933,7 @@ DialogBoxMorph.prototype.render = function (ctx) {
     ctx.beginPath();
     this.outlinePathTitle(
         ctx,
-        isFlat ? 0 : this.corner
+        this.corner
     );
     ctx.closePath();
     ctx.fill();
@@ -2946,7 +2944,7 @@ DialogBoxMorph.prototype.render = function (ctx) {
     ctx.beginPath();
     this.outlinePathBody(
         ctx,
-        isFlat ? 0 : this.corner
+        this.corner
     );
     ctx.closePath();
     ctx.fill();
@@ -3609,7 +3607,7 @@ PianoMenuMorph.prototype.createItems = function () {
     this.children.forEach(m => m.destroy());
     this.children = [];
     if (!this.isListContents) {
-        this.edge = MorphicPreferences.isFlat ? 0 : 5;
+        this.edge = MorphicPreferences.isFlat ? 3 : 5;
         this.border = MorphicPreferences.isFlat ? 1 : 2;
     }
     this.color = WHITE;